Refractory of the project structure
