+++ app/src/main/java/Helper/DbConnector.java
@@ -1,12 +1,9 @@
-package Database;
+package Helper;
 
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.IOException;
 import java.sql.Connection;
 import java.sql.DriverManager;
-import java.sql.SQLException;
 import java.util.Objects;
 import java.util.Properties;
